bot_whatsapp/requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
requests
python-dotenv
